+++ ssc/cmod_windpower.cpp
@@ -143,9 +143,9 @@ winddata::winddata(var_data *data_table)
 		return;
 	}
 
-	double* rh = get_vector(data_table, "rh", &len);
+	float* rh = get_vector(data_table, "rh", &len);
 	if (rh != 0 && len == data.nrows() )
-		m_relativeHumidity = std::vector<double>(rh, rh+(int)len);
+		m_relativeHumidity = std::vector<float>(rh, rh+(int)len);
 	else m_relativeHumidity.clear();
 }
