+++ tcs/csp_solver_util.h
@@ -70,7 +70,7 @@ public:
 	class C_output
 	{
 	private:
-		double *mp_reporting_ts_array;
+		float *mp_reporting_ts_array;
 		size_t m_n_reporting_ts_array;			//[-] Length of allocated array
 		std::vector<double> mv_temp_outputs;
 
@@ -88,7 +88,7 @@ public:
 
 		void set_m_is_ts_weighted(int subts_weight_type);
 
-		void assign(double *p_reporting_ts_array, size_t n_reporting_ts_array);
+		void assign(float *p_reporting_ts_array, size_t n_reporting_ts_array);
 
 		void set_timestep_output(double output_value);
 
@@ -126,7 +126,7 @@ public:
 
 	void construct(const S_output_info *output_info);
 
-	bool assign(int index, double *p_reporting_ts_array, size_t n_reporting_ts_array);
+	bool assign(int index, float *p_reporting_ts_array, size_t n_reporting_ts_array);
 
 	void send_to_reporting_ts_array(double report_time_start,
 		const std::vector<double> & v_temp_ts_time_end, double report_time_end);
