+++ test/shared_test/lib_weatherfile_test.cpp
@@ -194,9 +194,9 @@ protected:
 	void SetUp(){
 		// set-up input var_table: vd contains only zeros and time will be 1, 2, 3 
 		vt = new var_table;
-		double empty[8760] = { 0 };
+		float empty[8760] = { 0 };
 		vd = var_data(empty, 8760);
-		double order[3] = { 1, 2, 3 };
+		float order[3] = { 1, 2, 3 };
 		time = var_data(order, 3);
 		weatherdataTest::SetUp();
 		
@@ -274,9 +274,9 @@ protected:
 	void SetUp(){
 		// set-up input var_table
 		vt = new var_table;
-		double empty[9999] = { 0 };
+		float empty[9999] = { 0 };
 		vd = var_data(empty, 9999);
-		double order[3] = { 1, 2, 3 };
+		float order[3] = { 1, 2, 3 };
 		time = var_data(order, 3);
 		weatherdataTest::SetUp();
 	}
@@ -299,7 +299,7 @@ TEST_F(Data9999CaseWeatherData, initTest2_lib_weatherfile){
 }
 
 TEST_F(Data9999CaseWeatherData, readTest2_lib_weatherfile){
-	double wrong_length[1000] = { 0 };
+	float wrong_length[1000] = { 0 };
 	var_data vd_err = var_data(wrong_length, 1000);
 	input->table.unassign("dn");
 	input->table.assign("dn", vd_err);
